scalar: use 'git bugreport --diagnose' in 'scalar diagnose'
Replace implementation of 'scalar diagnose' with an internal invocation of
'git bugreport --diagnose --no-report'. The '--diagnose' option of 'git
bugreport' was implemented to mirror what 'scalar diagnose' does, taking
most of its code directly from 'scalar.c'. Remove the now-duplicate code in
'scalar.c' and have 'scalar diagnose' call 'git bugreport' to create the
diagnostics archive.

This introduces two (minor) changes to the output of 'scalar diagnose':
changing "Enlistment root" to "Repository root" in 'diagnostics.log'
("enlistment root" was inaccurate anyway, as the reported path always
pointed to the root of the repository), and changing the prefix of the zip
archive from 'scalar_' to 'git-diagnostics-'.

Signed-off-by: Victoria Dye <vdye@github.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>